fix(port): clear the compiler-specific spelling debt blocking the POSIX headless engine composition (ki-vuj)
Resumes ki-vuj after ki-0ahc (M1 exit) landed. Fresh headless-dedicated
configuration on linux (KISAK_BUILD_DEDICATED=ON, KISAK_DEDI_HEADLESS=ON,
gcc 16 -std=c++20) went from 1,191 errors at master 53652090 to 712, of
which 412 are the M4/M5 frozen-layout static_asserts and ~300 fall into
four documented out-of-scope buckets. Every edit below expands to
token-identical code on MSVC x86, preserving the M1 byte-parity invariant.

Spelling/compat fixes, all verified against the local portable suite
(207/210; the 3 failures - abi-sizeof-debt-tripwire,
abi-sizeof-scanner-fixture, security-source-regressions - are identical
on pristine master 53652090, verified in a clean worktree):

- src/bgame/bg_local.h: raw __declspec(align(8)) -> KISAK_ALIGNAS(8) for
  animation_s and animScriptData_t, and drop the meaningless 'const' on
  the shellshock_parms_t definition. The raw spellings broke the struct
  parses and cascaded ~250 'was not declared' errors per TU set.
- src/server_mp/server_mp.h: same KISAK_ALIGNAS(128) fix for
  serverStatic_t (unblocks the serverStatic_t/svs cascade).
- src/qcommon/msg_mp.cpp, src/qcommon/sv_msg_write_mp.cpp: 31 case
  labels spelled 0xFFFFFFxx on int switches; conformant compilers reject
  the narrowing while MSVC truncates. static_cast<int> keeps the
  identical bit pattern and hex greppability.
- src/universal/surfaceflags.h: (int)0x80000000 initializer casts for
  the nodrop/portal rows (same narrowing class).
- src/universal/q_shared.h GNUC branch: define _vsnprintf/_snprintf to
  the standard names; MSVC keeps its own spellings.
- include-order independence: graph.h, com_fileaccess.h,
  collision_trimesh_KISAK.h, physics/ode/util.h, buildnumber.cpp now
  include the dependency-free <universal/platform_compat.h> leaf so
  __cdecl is defined when the header is reached before q_shared.h.
- deps/ode/collision.h, deps/ode/objects.h: enum bases and members
  spelled with MSVC __int32/__int16 keywords -> exact-width <stdint.h>
  types; drop the redundant __cdecl from the KISAK-added
  dGetContactJointGroup declaration (default convention in extern 'C').
- src/physics/ode/collision_trimesh_KISAK.h: uint16_t spelling.
- src/qcommon/com_profilemapload.cpp: KISAK_PROFILE_TICKS() shim
  (__rdtsc / __builtin_ia32_rdtsc / monotonic fallback for arm64).
- src/sound/snd_public.h: guard the Win32-only SND_SetHWND hook.
- src/universal/com_sndalias_load_obj.cpp: file-local Com_IsValidName
  BOOL -> int (identical width and truth semantics).
- scripts/pre_build.cmake: POSIX target-dependency branch (Threads, DL,
  UNIX definitions, headless-no-Steam guard) replacing the
  unconditional FATAL_ERROR; scripts/increment_build.sh: +x bit.

CMakeLists.txt engine gate intentionally UNCHANGED and still in force:
the remaining 712 errors are 412 M4/M5 frozen-layout static_asserts
(bgame/qcommon/game/physics retail ABI mirrors) plus ~300 in four
deferred buckets: ~103 d3d9.h leaks into shared compositions through
gfx_d3d headers (xanim.h, xmodel.h, snd_public.h, fx_system.h,
cg_local_mp.h, direct includes in db_load.cpp/db_registry.cpp - M8/RHI-
scale type surgery), ~56 Win32 API surface via win32/win_local.h +
win_net.h + com_playerprofile.cpp (pending M3 sockets/system-info
services), ~17 raw-width pointer-into-int decompile patterns
(dvar.cpp/q_shared.cpp Info_Validate - M4-class decompile repair with
byte-parity discipline), and ~4 MSVC va_list artifacts in q_parse.cpp.
The linux-amd64-mp preset builds only the portable test suite; no
macos-arm64 preset exists yet, so the bead's recorded build/test
commands cannot exercise engine targets as written.

+++ src/qcommon/sv_msg_write_mp.cpp
@@ -1624,7 +1624,7 @@ void __cdecl MSG_WriteDeltaField(
     {
         switch (field->bits)
         {
-        case 0xFFFFFFA7:
+        case static_cast<int>(0xFFFFFFA7):
             fullFloat = *(float *)toF;
             trunc = (int)fullFloat;
             oldFloat = *(float *)fromF;
@@ -1646,11 +1646,11 @@ void __cdecl MSG_WriteDeltaField(
                 MSG_WriteByte(msg, trunc >> 5);
             }
             goto LABEL_103;
-        case 0xFFFFFFA8:
+        case static_cast<int>(0xFFFFFFA8):
             SV_PacketDataIsLargeFloat(snapInfo->clientNum, msg);
             MSG_WriteLong(msg, *fromF ^ *toF);
             goto LABEL_103;
-        case 0xFFFFFF9D:
+        case static_cast<int>(0xFFFFFF9D):
             fullFloat = *(float *)toF;
             trunc = (int)fullFloat;
             oldFloat = *(float *)fromF;
@@ -1690,7 +1690,7 @@ void __cdecl MSG_WriteDeltaField(
                     (int)(*(float *)toF + 2048.0),
                     4096);
             goto LABEL_103;
-        case 0xFFFFFF9C:
+        case static_cast<int>(0xFFFFFF9C):
             if (*toF)
             {
                 SV_PacketDataIsOverhead(snapInfo->clientNum, msg);
@@ -1705,13 +1705,13 @@ void __cdecl MSG_WriteDeltaField(
                 MSG_WriteBit0(msg);
             }
             goto LABEL_103;
-        case 0xFFFFFFA9:
+        case static_cast<int>(0xFFFFFFA9):
             SV_PacketDataIsOverhead(snapInfo->clientNum, msg);
             fullFloat = *(float *)toF;
             SV_PacketDataIsSmallAngle(snapInfo->clientNum, msg);
             MSG_WriteAngle16(msg, fullFloat);
             goto LABEL_103;
-        case 0xFFFFFFAA:
+        case static_cast<int>(0xFFFFFFAA):
             fullFloat = *(float *)toF;
             v16 = (fullFloat - 1.399999976158142) * 10.0f;
             trunc = SnapFloatToInt(v16);            
@@ -1727,23 +1727,23 @@ void __cdecl MSG_WriteDeltaField(
         {
             switch (field->bits)
             {
-            case 0xFFFFFF9F:
+            case static_cast<int>(0xFFFFFF9F):
                 MSG_WriteDeltaTime(snapInfo->clientNum, msg, time, *toF);
                 break;
-            case 0xFFFFFF9E:
+            case static_cast<int>(0xFFFFFF9E):
                 MSG_Write24BitFlag(snapInfo->clientNum, msg, *fromF, *toF);
                 break;
-            case 0xFFFFFFA0:
+            case static_cast<int>(0xFFFFFFA0):
                 MSG_WriteGroundEntityNum(snapInfo->clientNum, msg, *toF);
                 break;
-            case 0xFFFFFFA2:
+            case static_cast<int>(0xFFFFFFA2):
                 MSG_WriteEventNum(snapInfo->clientNum, msg, *toF);
                 break;
-            case 0xFFFFFFA3:
+            case static_cast<int>(0xFFFFFFA3):
                 MSG_WriteEventParam(snapInfo->clientNum, msg, *toF);
                 break;
-            case 0xFFFFFFA4:
-            case 0xFFFFFFA5:
+            case static_cast<int>(0xFFFFFFA4):
+            case static_cast<int>(0xFFFFFFA5):
                 fullFloat = *(float *)toF;
                 oldValue = *(float *)fromF;
                 //HIDWORD(v13) = msg;
@@ -1751,7 +1751,7 @@ void __cdecl MSG_WriteDeltaField(
                 //MSG_WriteOriginFloat(v13, field->bits, fullFloat, oldValue);
                 MSG_WriteOriginFloat(snapInfo->clientNum, msg, field->bits, fullFloat, oldValue);
                 break;
-            case 0xFFFFFFA6:
+            case static_cast<int>(0xFFFFFFA6):
                 fullFloat = *(float *)toF;
                 oldValue = *(float *)fromF;
                 //HIDWORD(v14) = msg;
@@ -1759,7 +1759,7 @@ void __cdecl MSG_WriteDeltaField(
                 //MSG_WriteOriginZFloat(v14, fullFloat, oldValue);
                 MSG_WriteOriginZFloat(snapInfo->clientNum, msg, fullFloat, oldValue);
                 break;
-            case 0xFFFFFFA1:
+            case static_cast<int>(0xFFFFFFA1):
                 value = *toF;
                 if (sv_debugPacketContents->current.enabled)
                     Com_Printf(16, "Sending %i as playerstate timer value (%ims granularity)\n", value, 100);
